+++ scripts/Muon/GUI/Common/seq_fitting_tab_widget/seq_fitting_tab_presenter.py
@@ -145,19 +145,10 @@ class SeqFittingTabPresenter(object):
             self._update_parameter_values_in_fitting_model_for_row(row)
 
     def validate_sequential_fit(self, workspace_names):
-        if self.model.get_active_fit_function() is None or len(workspace_names) == 0:
-            self.view.warning_popup("No data or fit function selected for fitting.")
-            return False
-        else:
-            return self._check_tf_asymmetry_compliance(self._flatten_workspace_names(workspace_names))
-
-    def _check_tf_asymmetry_compliance(self, workspace_names):
-        tf_compliant, non_compliant_names = self.model.check_datasets_are_tf_asymmetry_compliant(workspace_names)
-        if self.model.tf_asymmetry_mode and not tf_compliant:
-            self.view.warning_popup(f"Only Groups can be fitted in TF Asymmetry mode. Please unselect the following "
-                                    f"Pairs/Diffs in the grouping tab: {non_compliant_names}")
-            return False
-        return True
+        message = self.model.validate_sequential_fit(workspace_names)
+        if message != "":
+            self.view.warning_popup(message)
+        return message == ""
 
     @staticmethod
     def _flatten_workspace_names(workspaces: list) -> list:
